INC-41 refactor: put the logic into a hook
