+++ shared/assets/api/index.ts
@@ -1,1 +1,2 @@
 export * from "./base-api";
+export * from "./profile";
