+++ shared/assets/api/profile/index.ts
@@ -1,0 +1,5 @@
+export {
+  useDeleteUserPhotoMutation,
+  useFillOutProfileMutation,
+  useUploadUserPhotoMutation,
+} from "./profile-api";
