+++ shared/assets/helpers/index.ts
@@ -1,2 +1,3 @@
 export { checkPhoto } from "./checkPhoto";
 export { convertFileToBase64 } from "./convertFileToBase64";
+export { handleErrorResponse } from "./handleErrorResponse";
