+++ widgets/general/hook/index.ts
@@ -1,0 +1,2 @@
+export { useProfileForm } from "./useProfileForm";
+export { useUpdateAvatar } from "./useUpdateAvatar";
